requirements.txt
discord.py
chai-guanaco
html2image
beautifulsoup4